kitchensink updates
